requirements.txt
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
seaborn>=0.12
requests>=2.28
openpyxl>=3.1
pycausalimpact>=0.1.1
scipy>=1.10
statsmodels>=0.14
